Custom Makefile
Previously we were relying on rake-compiler to do everything and
not actually generating a shared library. This commit changes things
to use a custom Makefile that will generate the shared library as
well as the native tests.

It also moves a bunch of stuff around and attempts to simplify our
setup. I don't love that we still are using rake + make, as it feels
redundant, but I'm going to leave it as is for now. I think ideally
we'd just use rake, since then the whole setup would be simpler.