Lex with the parser
There is a lot of state that needs to be tracked that creates an
interdependence between the lexer and the parser. As such, we need
to have YARP.lex actually call yp_parse instead of looking at each
token one at a time.

Doing this means that the lexer now has access to things like the
current scope, which allows us to tell the difference between
things like divisions and regular expressions. That has been fixed
by this commit as well.

This commit does delete some tests that we had been using in order
to lex files individually, because it means that we now need to
properly parse them in order for it to work.