auto-generate specs based on parser testsuite
Introduces `bin/specs`, a program to hook into the parser testsuite to
auto-generate specs for the YARP parser. The specs are saved in a format
that is easy for other parser implementations to consume in order to
validate compatibility with YARP.

The spec generator hooks into `YARP#parse` to capture all of the source
code used in the parser tests.