Results for lexing
Previously, we only returned an array of tokens when we lexed. This
could be misleading, since the parser could have gathered up errors
and warnings without it being known to the consumer. Now when we
call lex or lex_compat, we return a ParseResult that contains the
tokens as well as errors, warnings, and comments.