Fixes for rescues
Previously, when we were comparing local variables we were only
comparing one of the lengths, so if you had a local like

    foobar

and then encountered an identifier like

    foo

it would think it was a local variable. This is now fixed.

Additionally, there was a weird state issue with rescue exception
variables where they could end up a => after a newline, but that's
been fixed as well.