New testing strategy
Previously we had a vendored version of ruby/spec that we checked
against. Now that we're further along, we want to test against more
targets.

This changes it so that rake lex knows how to download multiple
targets to test against. Each target keeps track of files that it
should exclude (files that are known to have syntax errors) as well
as files that should be excluded because they are not yet
supported. You can turn on the percentages with TODOS=1.