Remove clang-format
clang-format was added back when all of our code was in the ext
directory and has only been running on that directory since. I didn't
notice that we hadn't had it running for a while on src.

I spent a long time today trying to get a good config together that
would match our style, but it appears that none of them really work.
In particular there are two options that are really difficult to go
without.

The first is that there appears to be no way to have a newline between
the modifiers of a function and the function name. This is explicitly
the style in the CRuby codebase and I'd really like to match it.

The other is that clang-format appears to entirely mangle C-style
initializers. To the point that it's actually difficult to read.
There appears to be no formatting options to control this, and I don't
want to spend any more time than I already have.

I will happily accept a patch with a formatter that fixes those two
things, but until then I'm turning it off.