First string content being split on $#
In Ripper.lex, two string content nodes are created right after
another if the string contains $#, even if it doesn't result in
interpolation. This doesn't always happen though, and depends on
both state and other content in the string.

Replicating this behavior exactly in YARP would require more effort
than it's worth. Instead, we're going to just change this in our
own ripper lex function since we have semantically equivalent tokens
and no one is actually relying on this behavior.