Provide fixed-width to PP.pp for ParseTest
Because a string buffer is used rather than an IO device with a tty, PP
determines the width from its environment.  If `ENV["COLUMNS"]` was set
to anything besides `"80"`, then the snapshots will not match the `pp`
output.  This will happen when the shell exports the `COLUMNS` variable.

Additionally, the pretty printed line breaks in SourceFileNode must be
normalized or else long and short $CWD paths will result in different
output.  An alternative approach might be to normalize the nodes
directly, _prior_ to pretty printing.

Fixes #727