Allow generic option collection in templating parameters
This is a transitional PR which allows extra attributes of
parameters from config.yml to get saved as 'options' to the
parameter we interact with when generating code via ERB.

Why? In my current branch I have marked syntactical tokens that
execution is uninterested in serializing (and then loading into an
AST) as 'exclude: execute'.  My templates for serialize.c, Nodes.java,
and Loader.java then will skip those elements altogether. I can think
of other useful attributes we can put in to help with generation.