Lex against top 100 gems as part of CI
This commit introduces the lex:topgems rake task, and adds it to CI.
The rake task lexes against the top 100 gems, and compares their
failures to the known failures in tasks/top-100-gems.yml. If they
are not equivalent, the rake task will fail.

This will ensure we have no regressions when lexing the top 100 gems