Store serialized length for def nodes to support laziness
Implementations (JRuby/TruffleRuby) want to be able to lazily
load method definitions.  We store the length of the serialized
record so we can stash away that definition without processing it.
Once the method is called we process that serialized section.

length is defined as the serialized length including the entire
method definition minus the type byte.

This commit does not make either impl lazy but it does save length
so that a future commit can make it lazy.

The implementation is a little odd.  All nodes write:
{type, start_offset, end_offset} at the front of the node.  This
PR for def nodes then puts length after those 3 fields.  To
implement laziness we still have to read the initial location
fields before we get the length field.

The nature of the switch for construction makes moving this up
above the locations undesirable as we would need to emit
reading start/end into every case in the big node switch.  The
tradeoff of just reading locations to get to length feels reasonable.

serialize.rb just skips the field as there is no request for this
behavior there.