Rename fixture, snapshot files to have `.txt` endings
The fixture files are sometimes invalid Ruby code (for instance
a series of `break`s in a row). The snapshot files are not
Ruby files. We change their suffixes to be `.txt` to reflect this,
and also because there are test suites in ruby/ruby which will
assume that files ending in `.rb` are valid Ruby files.