Profiling and unicode work
This commit adds two things. The first is a build/profile make
target which is a small script that walks through a directory
recursively and parses every Ruby file that it finds. I use this
in combination with dappprof and dtrace to do some profiling to
determine where we're spending the most time.

The second thing this adds is a complete refactor of our UTF-8
decoder. Because UTF-8 is the default encoding and we need to know
if characters are alphanumeric or not very commonly, it's
important that this path is very fast. Previously we were always
decoding the unicode codepoint and then comparing it to a list of
known values, but that's incredibly inefficient when the character
is < 128 (i.e., almost all Ruby code). Instead we can use a lookup
table like we do for all the other encodings.

The UTF-8 decoding is actually significantly faster as well for the
few files that actually do need it. It now uses Bjoern Hoehrmann's
DFA-based UTF-8 decoder (modified to fit our use case).