Start using autotools for function detection
This commit introduces autotools for detecting whether or not some
functions exist.  The benefit of using autotools is that we don't need
to write our own feature detection C files, and that we'll use the same
`HAVE_*` macros that Ruby uses.  That means we can copy our C files in
to Ruby and piggyback off Ruby's configure scripts when Ruby is built.