next_newline and inline yp_strnstr
Previously our only use of memchr within yarp.c was to skip to the
next newline. We were also always casting to a size_t, which was
annoying and error prone in case it underflows.

Our only use of yp_stnstr was to find the encoding key within a
magic comment. We can inline it for better performance.