Fix use-after-free with lex modes
Lex modes can get popped when you lex the next token. If you want
to reference the previous value, you need to copy down the lex mode
before you lex the next token. We never ran into this before because
we keep around the previous next modes in a pre-allocated array
until we get up to 5.