Test symbol visibility in the static archive and shared objects
The tests check that the static archive librubyparser.a hides all of
the YARP functions, to ensure the symbols do not end up externally
visible in a dynamic symbol table and therefore not susceptible to
being replaced by the dynamic linker. We expect something like:

```
$ objdump --syms build/librubyparser.a | grep " yp_parse$"
000000000001dbf0 g     F .text	0000000000000009 .hidden yp_parse
```

The `g` means the symbol is global, but `.hidden` means that the
symbol should not be made external if linked into a shared
object (like `yarp.so`).

The tests also check that the shared object librubyparser.so contains
global versions of the public methods like `yp_parse`:

```
$ nm build/librubyparser.so | grep " yp_parse$"
0000000000021a50 T yp_parse
```

Finally, the tests verify that there is exactly one dynamic symbol in
`yarp.so`, which is the initializing function `Init_yarp`. NOTE that
we add `-fvisibility=hidden` to the C extension compiler flags to
ensure this test passes.