Switch compile-time default to NOT export symbols on non-windows
This reverses the convention introduced in #1069, and requires
`YP_EXPORT_SYMBOLS` to be defined at compile-time when building a
shared library.

See related #1111 which describes a symbol leakage problem in Ruby
3.3 (ruby head) that this commit, once merged upstream into Ruby,
should address.

Note that the tests introduced in a previous commit ensure that the
static archive and shared objects are all being created correctly
after this change.

Co-authored-by: Aaron Patterson <tenderlove@ruby-lang.org>