Eliminate warning: '_XOPEN_SOURCE' macro redefined
This commit eliminates the _XOPEN_SOURCE redefined warning.

YARP has 3 different ways to be compiled:

1. As part of CRuby
2. As a C extension gem
3. As a library to be linked against

CRuby defines XOPEN_SOURCE as well as HAVE_MMAP etc as part of its
configuration process.  In other words, when you run `./configure` in
Ruby on a target system, it will check for mmap etc and put those macros
in ruby/config.h.  When we mirror YARP to Ruby, we rewrite
`yarp/config.h` to simply include `ruby/config.h` so that YARP will use
the same HAVE_* macros

When C extensions are compiled, Ruby will pass compilation flags to the
C extension that is being compiled.  One of those compilations flags is
`-D_XOPEN_SOURCE`.  This clashes with the header file we generate with
via the M4 scripts.  Actually this is surfacing a bug because we
were generating `yarp/config.h` on the system building the YARP gem and
*not* on the target system compiling the gem.  This means that if we
build the gem on a system that supports mmap, but someone installs the
gem on a system that doesn't have mmap, there would be a bug.  To handle
this correctly, we _must_ test for the existence of mmap on the target
system at gem compile time.  To accommodate this, I've added
`have_func("mmap")` etc to `ext/yarp/extconf.rb`.  Those definitions are
passed to the compiler at gem compile time.  The gem is still compiled
with _XOPEN_SOURCE enabled because as I mentioned earlier: Ruby will
automatically pass that flag (as Ruby is compiled with that flag).

Finally, I changed the M4 scripts to pass XOPEN_SOURCE as well as
HAVE_MMAP, etc via the command line as well.  This means when we build
the library, the library will have XOPEN_SOURCE defined, but it doesn't
conflict with Ruby's XOPEN_SOURCE definition.