Remove recursion in rb_ast_new
Currently when we convert from the C structs to the Ruby objects we
recurse down the tree and create each node. This means we're limited
in the size of the tree that we can convert by the system stack
limit.

Instead, this commits changes the behavior to build up a stack of
nodes, and store their created value as we go. Effectively this is
an adaptation of the shunting yard algorithm but for creating Ruby
objects.

This change means we can now convert much larger trees into Ruby
objects without worrying about stack size.