Add filepath to yp_parse_serialize metadata
This adds the ability to pass the filepath through the
yp_parse_serialize function in the metadata. It also adds
documentation about this new API and fixes a couple of places that
were calling it with the old signature. Finally, it adds a specific
test to ensure yp_parse_serialize doesn't break.