Add UTF8-MAC encoding
Because we only need a subset of functionality, we should be able
to reuse the UTF-8 encoding here. (I think?)

I checked by running:

```ruby
(0...0x10000).each do |codepoint|
  next if codepoint >= 0xD800 .. codepoint <= 0xDFFF

  utf = codepoint.chr(Encoding::UTF_8)
  mac = codepoint.chr(Encoding::UTF8_MAC)

  if utf.match?(/[[:alpha:]]/) != mac.match?(/[[:alpha:]]/)
    puts "0x#{codepoint.to_s(16).upcase}"
  end

  if utf.match?(/[[:alnum:]]/) != mac.match?(/[[:alnum:]]/)
    puts "0x#{codepoint.to_s(16).upcase}"
  end

  if utf.match?(/[[:upper:]]/) != mac.match?(/[[:upper:]]/)
    puts "0x#{codepoint.to_s(16).upcase}"
  end
end
```

Since nothing was printed, it means the character properties that
we care about match. I think this means we can reuse it.