Check locals
We want to ensure that we have the same set of locals for any given
scope as CRuby does so that we can compile them correctly.

Currently we need to skip "special" locals like the keyword
arguments hash and numbered parameters, but this gets us to a place
we can iterate.