Fix #1793.  offsets off of end of source should not assert.
There are cases where prism will run into a syntax error and advance to
at least 1 byte past the end of source (beyond just regexps like as described
in #1793).  There are many places this can happen so this PR fixes this
issue by adjusting the offset we are trying to determine the line for to
the last valid byteoffset in the source.