Fix template generation of Loader.java for JRuby
The original work nearly worked but 3 fields needed to be accessed
by the extended type JRuby needs to create to properly have
`getName()` return RubySymbol instead of j.l.String.

This PR additional removes an unused field when generating JRuby version
of Loader.java because it is only needed for `getName()` when it
returns j.l.String.