Fix bin/lex when encountering mismatch at EOF.
When encountering a situation where there is a mismatch at the very end
of a file, and when `Prism.lex` produces a shorter output than
`.lex_compat` and/or `.lex_ripper`, `bin/lex` itself fails with
`NoMethodError` by calling `[0]` on `new` which is `nil`.
This prevents outputting the actual mismatch.