Improve to handle unterminated strings
Fix #1946

This fixes to set an error position for unterminated strings to the
opening delimiters. Previously, the error position was set to the end
of the delimiter.

The same fix applies to other string-like literals.

Additionally, this fixes #1946; that is, it adds the last part of the
string even though the string literal does not terminate.