Make Sorbet typechecking step use `typed: true` by default
This change makes the Sorbet typechecking step use `typed: true` by
default for all files except the ones we explicitly typed override back
to `typed: false`.

We also extend the Sorbet config to suppress a few classes of errors
that are not essential and would be hard to fix without inline type annotations.